allow for disabling of rendergroup split tests
